ALERT: Added unknown source basic.elf
Add basic.elf from

https://github.com/ultraembedded/riscv/blob/master/isa_sim/images/basic.elf

which doesn't seem to have a corresponding source file to make sure
there is no license issue. Next to this, in the original repository,
there is a linux.elf. The readme of the original repo says:

    "There are two example pre-compiled ELFs provided, one which is a
    basic machine mode only test program, and one which boots Linux
    (modified 4.19 compiled for RV32IM)."

see https://github.com/ultraembedded/riscv/blob/master/isa_sim/README.md#usage

I'm not knowingly breaking the GPL of the Linux kernel, but even copying
basic.elf (and generating the disassembly and raw binary from it) is not
wihtout possible issues, yet the LICENSE says is OK, so I hope it was
made from a known source that was allowed distribution.

In either case, I should replace this file as soon as possible with a
self generated binary or some image of some obviously
legal-to-distribute binary such as a FreeRTOS image for one of the
RISC-V microcontrollers.

Since Moaca does not link to the basic.elf binary, there is no real
risk, as another reference binary to run in the emulator should be easy
to obtain from a known working binary like FreeRTOS.

Also generated raw and disassembled equivalents, as is easier to work
with those.

files obtained from .elf with:

  Disassembly:

    llvm-objdump-13 --arch=riscv32 -d riscv/isa_sim/images/basic.elf > basic.disassembled

  Raw memory dump of the elf:

    llvm-objcopy-13 -O binary riscv/isa_sim/images/basic.elf basic.binary

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>